Mesh refactoring - move renderer away
