+++ src/mesh/Mesh.js
@@ -1,6 +1,4 @@
 import * as core from '../core';
-import glCore from 'pixi-gl-core';
-import Shader from './webgl/MeshShader';
 
 const tempPoint = new core.Point();
 const tempPolygon = new core.Polygon();
@@ -59,11 +57,17 @@ export default class Mesh extends core.Container
         this.indices = indices || new Uint16Array([0, 1, 3, 2]);
 
         /**
-         * Whether the Mesh is dirty or not
+         * Version of mesh uvs are dirty or not
          *
          * @member {number}
          */
         this.dirty = 0;
+
+        /**
+         * Version of mesh indices
+         *
+         * @member {number}
+         */
         this.indexDirty = 0;
 
         /**
@@ -122,62 +126,8 @@ export default class Mesh extends core.Container
      */
     _renderWebGL(renderer)
     {
-        // get rid of any thing that may be batching.
-        renderer.flush();
-
-        //  renderer.plugins.mesh.render(this);
-        const gl = renderer.gl;
-        let glData = this._glDatas[renderer.CONTEXT_UID];
-
-        if (!glData)
-        {
-            glData = {
-                shader: new Shader(gl),
-                vertexBuffer: glCore.GLBuffer.createVertexBuffer(gl, this.vertices, gl.STREAM_DRAW),
-                uvBuffer: glCore.GLBuffer.createVertexBuffer(gl, this.uvs, gl.STREAM_DRAW),
-                indexBuffer: glCore.GLBuffer.createIndexBuffer(gl, this.indices, gl.STATIC_DRAW),
-                // build the vao object that will render..
-                vao: new glCore.VertexArrayObject(gl),
-                dirty: this.dirty,
-                indexDirty: this.indexDirty,
-            };
-
-            // build the vao object that will render..
-            glData.vao = new glCore.VertexArrayObject(gl)
-                .addIndex(glData.indexBuffer)
-                .addAttribute(glData.vertexBuffer, glData.shader.attributes.aVertexPosition, gl.FLOAT, false, 2 * 4, 0)
-                .addAttribute(glData.uvBuffer, glData.shader.attributes.aTextureCoord, gl.FLOAT, false, 2 * 4, 0);
-
-            this._glDatas[renderer.CONTEXT_UID] = glData;
-        }
-
-        if (this.dirty !== glData.dirty)
-        {
-            glData.dirty = this.dirty;
-            glData.uvBuffer.upload();
-        }
-
-        if (this.indexDirty !== glData.indexDirty)
-        {
-            glData.indexDirty = this.indexDirty;
-            glData.indexBuffer.upload();
-        }
-
-        glData.vertexBuffer.upload();
-
-        renderer.bindShader(glData.shader);
-        renderer.bindTexture(this._texture, 0);
-        renderer.state.setBlendMode(this.blendMode);
-
-        glData.shader.uniforms.translationMatrix = this.worldTransform.toArray(true);
-        glData.shader.uniforms.alpha = this.worldAlpha;
-        glData.shader.uniforms.tint = this.tintRgb;
-
-        const drawMode = this.drawMode === Mesh.DRAW_MODES.TRIANGLE_MESH ? gl.TRIANGLE_STRIP : gl.TRIANGLES;
-
-        glData.vao.bind()
-            .draw(drawMode, this.indices.length)
-            .unbind();
+        renderer.setObjectRenderer(renderer.plugins.mesh);
+        renderer.plugins.mesh.render(this);
     }
 
     /**
@@ -188,240 +138,7 @@ export default class Mesh extends core.Container
      */
     _renderCanvas(renderer)
     {
-        const context = renderer.context;
-
-        const transform = this.worldTransform;
-        const res = renderer.resolution;
-
-        if (renderer.roundPixels)
-        {
-            context.setTransform(
-                transform.a * res,
-                transform.b * res,
-                transform.c * res,
-                transform.d * res,
-                (transform.tx * res) | 0,
-                (transform.ty * res) | 0
-            );
-        }
-        else
-        {
-            context.setTransform(
-                transform.a * res,
-                transform.b * res,
-                transform.c * res,
-                transform.d * res,
-                transform.tx * res,
-                transform.ty * res
-            );
-        }
-
-        if (this.drawMode === Mesh.DRAW_MODES.TRIANGLE_MESH)
-        {
-            this._renderCanvasTriangleMesh(context);
-        }
-        else
-        {
-            this._renderCanvasTriangles(context);
-        }
-    }
-
-    /**
-     * Draws the object in Triangle Mesh mode using canvas
-     *
-     * @private
-     * @param {CanvasRenderingContext2D} context - The current drawing context
-     */
-    _renderCanvasTriangleMesh(context)
-    {
-        // draw triangles!!
-        const vertices = this.vertices;
-        const uvs = this.uvs;
-        const length = vertices.length / 2;
-
-        // this.count++;
-
-        for (let i = 0; i < length - 2; i++)
-        {
-            // draw some triangles!
-            const index = i * 2;
-
-            this._renderCanvasDrawTriangle(context, vertices, uvs, index, (index + 2), (index + 4));
-        }
-    }
-
-    /**
-     * Draws the object in triangle mode using canvas
-     *
-     * @private
-     * @param {CanvasRenderingContext2D} context - the current drawing context
-     */
-    _renderCanvasTriangles(context)
-    {
-        // draw triangles!!
-        const vertices = this.vertices;
-        const uvs = this.uvs;
-        const indices = this.indices;
-        const length = indices.length;
-        // this.count++;
-
-        for (let i = 0; i < length; i += 3)
-        {
-            // draw some triangles!
-            const index0 = indices[i] * 2;
-            const index1 = indices[i + 1] * 2;
-            const index2 = indices[i + 2] * 2;
-
-            this._renderCanvasDrawTriangle(context, vertices, uvs, index0, index1, index2);
-        }
-    }
-
-    /**
-     * Draws one of the triangles that form this Mesh
-     *
-     * @private
-     * @param {CanvasRenderingContext2D} context - the current drawing context
-     * @param {Float32Array} vertices - a reference to the vertices of the Mesh
-     * @param {Float32Array} uvs - a reference to the uvs of the Mesh
-     * @param {number} index0 - the index of the first vertex
-     * @param {number} index1 - the index of the second vertex
-     * @param {number} index2 - the index of the third vertex
-     */
-    _renderCanvasDrawTriangle(context, vertices, uvs, index0, index1, index2)
-    {
-        const base = this._texture.baseTexture;
-        const textureSource = base.source;
-        const textureWidth = base.width;
-        const textureHeight = base.height;
-
-        const u0 = uvs[index0] * base.width;
-        const u1 = uvs[index1] * base.width;
-        const u2 = uvs[index2] * base.width;
-        const v0 = uvs[index0 + 1] * base.height;
-        const v1 = uvs[index1 + 1] * base.height;
-        const v2 = uvs[index2 + 1] * base.height;
-
-        let x0 = vertices[index0];
-        let x1 = vertices[index1];
-        let x2 = vertices[index2];
-        let y0 = vertices[index0 + 1];
-        let y1 = vertices[index1 + 1];
-        let y2 = vertices[index2 + 1];
-
-        if (this.canvasPadding > 0)
-        {
-            const paddingX = this.canvasPadding / this.worldTransform.a;
-            const paddingY = this.canvasPadding / this.worldTransform.d;
-            const centerX = (x0 + x1 + x2) / 3;
-            const centerY = (y0 + y1 + y2) / 3;
-
-            let normX = x0 - centerX;
-            let normY = y0 - centerY;
-
-            let dist = Math.sqrt((normX * normX) + (normY * normY));
-
-            x0 = centerX + ((normX / dist) * (dist + paddingX));
-            y0 = centerY + ((normY / dist) * (dist + paddingY));
-
-            //
-
-            normX = x1 - centerX;
-            normY = y1 - centerY;
-
-            dist = Math.sqrt((normX * normX) + (normY * normY));
-            x1 = centerX + ((normX / dist) * (dist + paddingX));
-            y1 = centerY + ((normY / dist) * (dist + paddingY));
-
-            normX = x2 - centerX;
-            normY = y2 - centerY;
-
-            dist = Math.sqrt((normX * normX) + (normY * normY));
-            x2 = centerX + ((normX / dist) * (dist + paddingX));
-            y2 = centerY + ((normY / dist) * (dist + paddingY));
-        }
-
-        context.save();
-        context.beginPath();
-
-        context.moveTo(x0, y0);
-        context.lineTo(x1, y1);
-        context.lineTo(x2, y2);
-
-        context.closePath();
-
-        context.clip();
-
-        // Compute matrix transform
-        const delta = (u0 * v1) + (v0 * u2) + (u1 * v2) - (v1 * u2) - (v0 * u1) - (u0 * v2);
-        const deltaA = (x0 * v1) + (v0 * x2) + (x1 * v2) - (v1 * x2) - (v0 * x1) - (x0 * v2);
-        const deltaB = (u0 * x1) + (x0 * u2) + (u1 * x2) - (x1 * u2) - (x0 * u1) - (u0 * x2);
-        const deltaC = (u0 * v1 * x2) + (v0 * x1 * u2) + (x0 * u1 * v2) - (x0 * v1 * u2) - (v0 * u1 * x2) - (u0 * x1 * v2);
-        const deltaD = (y0 * v1) + (v0 * y2) + (y1 * v2) - (v1 * y2) - (v0 * y1) - (y0 * v2);
-        const deltaE = (u0 * y1) + (y0 * u2) + (u1 * y2) - (y1 * u2) - (y0 * u1) - (u0 * y2);
-        const deltaF = (u0 * v1 * y2) + (v0 * y1 * u2) + (y0 * u1 * v2) - (y0 * v1 * u2) - (v0 * u1 * y2) - (u0 * y1 * v2);
-
-        context.transform(
-            deltaA / delta,
-            deltaD / delta,
-            deltaB / delta,
-            deltaE / delta,
-            deltaC / delta,
-            deltaF / delta
-        );
-
-        context.drawImage(
-            textureSource,
-            0,
-            0,
-            textureWidth * base.resolution,
-            textureHeight * base.resolution,
-            0,
-            0,
-            textureWidth,
-            textureHeight
-        );
-
-        context.restore();
-    }
-
-    /**
-     * Renders a flat Mesh
-     *
-     * @private
-     * @param {PIXI.mesh.Mesh} mesh - The Mesh to render
-     */
-    renderMeshFlat(mesh)
-    {
-        const context = this.context;
-        const vertices = mesh.vertices;
-        const length = vertices.length / 2;
-
-        // this.count++;
-
-        context.beginPath();
-
-        for (let i = 1; i < length - 2; ++i)
-        {
-            // draw some triangles!
-            const index = i * 2;
-
-            const x0 = vertices[index];
-            const y0 = vertices[index + 1];
-
-            const x1 = vertices[index + 2];
-            const y1 = vertices[index + 3];
-
-            const x2 = vertices[index + 4];
-            const y2 = vertices[index + 5];
-
-            context.moveTo(x0, y0);
-            context.lineTo(x1, y1);
-            context.lineTo(x2, y2);
-        }
-
-        context.fillStyle = '#FF0000';
-        context.fill();
-        context.closePath();
+        renderer.plugins.mesh.render(this);
     }
 
     /**
