+++ src/mesh/index.js
@@ -2,7 +2,8 @@
  * @namespace PIXI.mesh
  */
 export { default as Mesh } from './Mesh';
+export { default as MeshRenderer } from './webgl/MeshRenderer';
+export { default as CanvasMeshRenderer } from './canvas/CanvasMeshRenderer';
 export { default as Plane } from './Plane';
 export { default as NineSlicePlane } from './NineSlicePlane';
 export { default as Rope } from './Rope';
-export { default as MeshShader } from './webgl/MeshShader';
